# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schema App - Todo",
        "description": "Simple Todo Restful API",
        "termsOfService": "https://www.domain.com",
        "contact": {
            "email": "hamedv90@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Todo"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "List of tasks",
                "description": "List of tasks",
                "operationId": "tasks_index",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by Category Id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortbyDueDate",
                        "in": "query",
                        "description": "Sort by due date",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Spring cleaning"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Get rid of unnecessary items!"
                                                    },
                                                    "due_date_at": {
                                                        "type": "string",
                                                        "example": "2023-03-16 19:14:50"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Personal"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tasks retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new tasks",
                "description": "Create a new tasks",
                "operationId": "tasks_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Find an apartment"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Coquitlam or Burnaby"
                                    },
                                    "due_date_at": {
                                        "type": "string",
                                        "format": "YYYY-MM-DD HH:MM:SS",
                                        "example": "2023-03-21 23:34:43"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Find an apartment"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Coquitlam or Burnaby"
                                                },
                                                "due_date_at": {
                                                    "type": "string",
                                                    "example": "2023-03-21 23:34:43"
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}